File restructure (#100)
* maths directory
* move fit classes to classes.py
