+++ gpfit/__init__.py
@@ -1,2 +1,4 @@
 "Package for GP-compatible data fitting"
 __version__ = "1.0.0"
+
+from .fit import fit
