+++ gpfit/maths/__init__.py
@@ -1,0 +1,1 @@
+"Modules containing functions that perform the mathematics of gpfit"
