+++ gpfit/maths/logsumexp.py
